requirements.txt
google-adk
ollama
pdf2image
pillow
pydantic
pytest
